lesson4_homework_0.5
