+++ lesson4/main.py
@@ -1,11 +1,11 @@
 from scrapy.crawler import CrawlerProcess
 from scrapy.settings import Settings
 from gbpars import settings
-from gbpars.spiders.avito import AvitoSpider
+from gbpars.spiders.youla import YoulaSpider
 
 if __name__ == "__main__":
     crawl_settings = Settings()
     crawl_settings.setmodule(settings)
     crawl_proc = CrawlerProcess(settings=crawl_settings)
-    crawl_proc.crawl(AvitoSpider)
+    crawl_proc.crawl(YoulaSpider)
     crawl_proc.start()
